refact: refactoring rootstack paramlist
