+++ src/screens/contentList/ContentList.tsx
@@ -33,7 +33,7 @@ import { PopupMenu } from '@/models/PopupMenu'
 import { ReportMenuType, ReportMenus } from '@/models/enums/ActionSheetType'
 import { ReportType } from '@/models/enums/ReportType'
 import { MainNavigationProp } from '@/navigations/MainNavigator'
-import { RootStackParamList } from '@/navigations/RootStack'
+import { RootStackParamList } from '@/navigations/RootStackParamList'
 import { isCloseToBottom } from '@/services/InfiniteService'
 import { getActionSheetTintColor } from '@/services/StyleService'
 import { CategoryState, CommunityCategoryState } from '@/state/CategoryState'
